+++ test/test_helper.rb
@@ -16,8 +16,10 @@ Rails.backtrace_cleaner.remove_silencers!
 Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each { |f| require f }
 
 def setup_test_fixture
-  @user = User.new(email: 'steve@example.com', password: 'secret', comments: 'my comment')
+  @user = User.create!(email: 'steve@example.com', password: 'secret', comments: 'my comment', terms: true)
   @builder = BootstrapForm::FormBuilder.new(:user, @user, self, {})
+  @user_new = User.new(email: 'steve@example.com', password: 'secret', comments: 'my comment')
+  @builder_new = BootstrapForm::FormBuilder.new(:user, @user_new, self, {})
   @horizontal_builder = BootstrapForm::FormBuilder.new(:user, @user, self, { layout: :horizontal, label_col: "col-sm-2", control_col: "col-sm-10" })
   I18n.backend.store_translations(:en, {
     activerecord: {