+++ test/dummy/db/migrate/20130912171202_add_preferences_to_user.rb
@@ -1,4 +1,4 @@
-class AddPreferencesToUser < ActiveRecord::Migration
+class AddPreferencesToUser < GenericMigration
   def change
     add_column :users, :preferences, :text
   end