+++ test/dummy/db/migrate/20140922133133_add_type_to_users.rb
@@ -1,4 +1,4 @@
-class AddTypeToUsers < ActiveRecord::Migration
+class AddTypeToUsers < GenericMigration
   def change
     add_column :users, :type, :string
   end