+++ CHANGELOG.md
@@ -5,6 +5,11 @@ Bugfixes:
 
 Features:
   - Your contribution here!
+  
+## [Pending Release][]
+  
+Bugfixes:
+  - Fix an issue that occurs when the output buffer of the view layer does not respond to "concat" (i.e. an Array).
 
 ## [2.7.0][] (2017-04-21)